+++ tests/testthat/test-pagination.R
@@ -1,11 +1,11 @@
-test_that("export_as_docx works with page_by", {
-
-})
-
-test_that("export_as_docx works with list of {rtables}' tables", {
-
-})
-
-test_that("export_as_docx works with list of {rtables}' and {rlistings}' objects", {
-
-})
+# test_that("export_as_docx works with page_by", {
+#
+# })
+#
+# test_that("export_as_docx works with list of {rtables}' tables", {
+#
+# })
+#
+# test_that("export_as_docx works with list of {rtables}' and {rlistings}' objects", {
+#
+# })