+++ NEWS.md
@@ -1,12 +1,11 @@
 ## rtables.officer 0.0.2.9003
 
-### New Features
  * Added option to flush pages when exporting to `.docx`. It is possible to put tables on separate pages by doing `export_as_docx(add_page_break = TRUE)`.
-
+ * Now it is possible to tranform and export also `rlistings` objects.
+ * Added separator lines and padding when present in the `rlistings` or `rtables` objects.
 
 ## rtables.officer 0.0.2
 
-### New Features
  * Experimental pagination is now possible in `tt_as_flextable()` and `export_as_docx()`.
  * Added handling of widths in `tt_as_flextable()`. Now it is possible to change column widths for `.docx` exports.
  * First version of `rtables.officer`, split from the `rtables` package.